Got output folder working and ability to save image in a new directory. Started metadata handler.

+++ src/main/java/org/lifesync/utilities/MediaCopier.java
@@ -13,6 +13,10 @@ public class MediaCopier {
    * @param destination The path of the copied file to create.
    */
   public void copy(final Path source, final Path destination) {
+    if (Files.exists(destination)) {
+      return;
+    }
+
     try {
       Files.copy(source, destination, StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.COPY_ATTRIBUTES);
     } catch (UnsupportedOperationException e) {
